fix: 진영별 잔여 기물 점수 기준으로 점수 계산을 수정
실제 장기 규칙에 알맞게 한 진영은 73.5점을 초 진영은 72점을 가진 상태로 초기화

+++ src/main/java/janggi/domain/ScoreStatus.java
@@ -1,4 +1,4 @@
 package janggi.domain;
 
-public record ScoreStatus(int choScore, int hanScore) {
+public record ScoreStatus(double choScore, double hanScore) {
 }
